requirements.txt
future~=0.18.2
Pillow~=9.0.1
paramiko~=2.9.3
pip~=21.3.1
wheel~=0.37.1
cryptography~=3.4.8
keyring~=23.5.0
setuptools==63.2.0
numpy
pyMapVBVD==0.5.2
opencv-python
matplotlib
scipy
scikit-image
joblib
nibabel
sklearn
numba
tqdm
pyqt6